# chore: avoid needless compression