Previously, grype could only import a compressed vulnerability database (the result of 'grype-db package'). However, now that grype can import an uncompressed database (the result of 'grype-db build' without needing to package), skip the costly compression step for databases that will not be distributed, such as those that will be reused on the same host in CI or built during local development.

## Performance improvements

building database during quality gate is ~5 min faster for alpine
```
# in tests/quality
# before
make build-db provider=alpine  713.32s user 129.58s system 124% cpu 11:15.60 total
# after
make build-db provider=alpine  251.56s user 122.78s system 110% cpu 5:38.57 total
```

building development db (alpine with nvd cache present) is ~5 min faster

```
# without changes (building alpine with nvd cached)
make update-db  746.56s user 115.34s system 131% cpu 10:55.49 total
# with changes (building alpine with nvd cached)
make update-db  255.56s user 108.75s system 124% cpu 4:53.20 total
```